add 012 upgrade option (#6)
./terrafmt upgrade012 path/to/test_files.go

Because terraform 0.12upgrade needs to be run on a directory and doesn't have a stdin option, it will write the blocks to a buffer, create a temp directory, write the buffer to a temp file there, run terraform 0.12upgrade on the temp directory, read the temp file back to the format block before it writes it to the real file.

I didn't implement the fmtcompat option into this. I didn't entirely understand the use case, but I can add it in if you need me to.